refactor(plugin-api): enhance chart data handling by adding new functions for generating pie and trend charts, and streamline imports for improved clarity
