+++ packages/plugin-awesome/src/charts.ts
@@ -1,38 +1,15 @@
-import {
-  ChartDataType,
-  ChartType,
-  type HistoryDataPoint,
-  type Statistic,
-  type TestResult,
-} from "@allurereport/core-api";
+import { ChartType } from "@allurereport/core-api";
 import {
   type AllureStore,
-  DEFAULT_CHART_HISTORY_LIMIT,
   type GeneratedChartData,
   type GeneratedChartsData,
-  type PieChartData,
-  type PieChartOptions,
   type PluginContext,
-  type TrendChartData,
-  type TrendChartOptions,
-  getPieChartData,
-  getSeverityTrendData,
-  getStatusTrendData,
+  generateTrendChart,
+  generatePieChart,
 } from "@allurereport/plugin-api";
 import { randomUUID } from "crypto";
 import type { AwesomeOptions } from "./model.js";
 import type { AwesomeDataWriter } from "./writer.js";
-
-const generatePieChart = (
-  options: PieChartOptions,
-  stores: {
-    statistic: Statistic;
-  },
-): PieChartData => {
-  const { statistic } = stores;
-
-  return getPieChartData(statistic, options);
-};
 
 export const generateCharts = async (
   options: AwesomeOptions,
@@ -74,26 +51,6 @@ export const generateCharts = async (
 
     return acc;
   }, {} as GeneratedChartsData);
-};
-
-export const generateTrendChart = (
-  options: TrendChartOptions,
-  stores: {
-    historyDataPoints: HistoryDataPoint[];
-    statistic: Statistic;
-    testResults: TestResult[];
-  },
-  context: PluginContext,
-): TrendChartData | undefined => {
-  const newOptions = { limit: DEFAULT_CHART_HISTORY_LIMIT, ...options };
-  const { dataType } = newOptions;
-  const { statistic, historyDataPoints, testResults } = stores;
-
-  if (dataType === ChartDataType.Status) {
-    return getStatusTrendData(statistic, context.reportName, historyDataPoints, newOptions);
-  } else if (dataType === ChartDataType.Severity) {
-    return getSeverityTrendData(testResults, context.reportName, historyDataPoints, newOptions);
-  }
 };
 
 export const generateAllCharts = async (
